requirements.txt
blinker==1.6.3
boto3==1.28.66
botocore==1.31.66
click==8.1.7
Flask==3.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.3
python-dateutil==2.8.2
python-dotenv==1.0.0
s3transfer==0.7.0
six==1.16.0
urllib3==2.0.7
Werkzeug==3.0.0
